Cleanup code

+++ DailyCodingProblem/Problem_28.h
@@ -1,14 +1,18 @@
 #pragma once
 /* MEDIUM
-Write an algorithm to justify text. Given a sequence of words and an integer line length k, return a list of strings which represents each line, fully justified.
+Write an algorithm to justify text. Given a sequence of words and an integer line length k, return a list of strings
+which represents each line, fully justified.
 
-More specifically, you should have as many words as possible in each line. There should be at least one space between each word. Pad extra spaces when necessary so that each line has exactly length k. Spaces should be distributed as equally as possible, with the extra spaces, if any, distributed starting from the left.
+More specifically, you should have as many words as possible in each line. There should be at least one space between
+each word. Pad extra spaces when necessary so that each line has exactly length k. Spaces should be distributed as
+equally as possible, with the extra spaces, if any, distributed starting from the left.
 
 If you can only fit one word on a line, then you should pad the right-hand side with spaces.
 
 Each word is guaranteed not to be longer than k.
 
-For example, given the list of words ["the", "quick", "brown", "fox", "jumps", "over", "the", "lazy", "dog"] and k = 16, you should return the following:
+For example, given the list of words ["the", "quick", "brown", "fox", "jumps", "over", "the", "lazy", "dog"] and k = 16,
+you should return the following:
 
 ["the  quick brown", # 1 extra space on the left
 "fox  jumps  over", # 2 extra spaces distributed evenly
@@ -24,7 +28,7 @@ std::vector<std::string> padWords( std::vector<std::string> words, unsigned int 
 
   while( !words.empty() )
   {
-    unsigned int             length = 0;
+    size_t                   length = 0;
     std::vector<std::string> line;
 
     while( length < k )
@@ -41,7 +45,7 @@ std::vector<std::string> padWords( std::vector<std::string> words, unsigned int 
         break;
     }
     length += line.size() - 1;
-    int addSpaces = k - length;
+    size_t addSpaces = k - length;
 
     while( addSpaces > 0 )
     {
